pyproject.toml
[project]
name = "stem-agent"
version = "0.1.0"
description = "A self-specializing AI agent framework inspired by biological stem cells"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=1.1.0",
    "langchain-core>=1.2.0",
    "langchain-openai>=0.3.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "truststore>=0.9.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
